+++ src/Estados/movimiento.cpp
@@ -6,7 +6,7 @@
 
 //VER CUANTO MOVER A LA UNIDAD SEGUN EL TIEMPO TRANSCURRIDO Y SU VELOCIDAD
 std::shared_ptr<Estado> Movimiento::actualizar(UnidadMovible *unidad,
-Mapa &mapa, double tiempo_transcurrido) {
+Mapa &mapa, int tiempo_transcurrido) {
 	std::list<std::pair<int, int>> camino = unidad->pedir_camino();
 	bool posicion_ocupada = mapa.esta_ocupada_coordenada(camino.front());
 	if (posicion_ocupada && camino.size() == 1) {
