+++ src/Estados/movimiento_para_atacar.h
@@ -18,7 +18,7 @@ class MovimientoParaAtacar: public Movimiento {
 
 		/*Metodo virtual que actualiza el estado.*/
 		virtual std::shared_ptr<Estado> actualizar(UnidadMovible *unidad,
-		Mapa &mapa, double tiempo_transcurrido);
+		Mapa &mapa, int tiempo_transcurrido);
 };
 
 #endif
