+++ src/Estados/movimiento_para_depositar.cpp
@@ -10,7 +10,7 @@ coordenada_a_volver(terreno) {
 }
 
 std::shared_ptr<Estado> MovimientoParaDepositar::actualizar(UnidadMovible 
-*unidad, Mapa &mapa, double tiempo_transcurrido) {
+*unidad, Mapa &mapa, int tiempo_transcurrido) {
 	std::list<std::pair<int, int>> camino = mapa.obtener_camino(
 	unidad->obtener_centro(), refineria_destino->obtener_centro());
 	unidad->asignar_nuevo_camino(camino);
