+++ src/Estados/muerte.cpp
@@ -1,7 +1,7 @@
 #include "muerte.h"
 
 std::shared_ptr<Estado> Muerte::actualizar(UnidadMovible *unidad,
-Mapa &mapa, double tiempo_transcurrido) {
+Mapa &mapa, int tiempo_transcurrido) {
 	nombre = "muerte";
 	unidad->serializar_mensaje_muerte();
 	std::vector<ObjetoDune*> unidades_afectadas = unidad->
